requirements.txt
langchain
langchain-openai
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pydantic
pypandoc
google-api-python-client
google-auth
